+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file documents the most important changes for each released version.
 
+## [v1.0.0]
+- Added ping and jitter options
+- Improved mangler API so that it can by stopped asynchronously
+- Added basic GUI
+
 ## [v0.1.1]
 - Switched GUI renderer to wgpu
 - Added automated publish